+++ CHANGELOG.md
@@ -9,7 +9,11 @@
 
 - Filesystem errors on some Windows environments when running coilpack install command
 - GraphQL fieldtype modifier registration bug
-- Use coding standard
+- GraphQL support for relationship fieldtype
+
+### Changed
+
+- Use Laravel coding standard
 - Improved typehints
 
 ## [0.0.1] - 2022-12-29